chore(server): refactor and unify tokens infrastructure tests
